+++ scripts/lint-fix.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env bash
+
+set -exu
+
+ruff check --fix ksuid
+ruff format ksuid